Migrate SAM Record POS/PNEXT/TLEN to int64 for long references (gap A14)
sam.Record.Pos/PNext/TLen are now int64 (htslib hts_pos_t), so SAM and CRAM
round-trip reference coordinates beyond 2^31 (htslib's longref.sam). BAM keeps
its 32-bit on-disk limit: the writer ERRORS cleanly for POS/PNEXT 0-based >
MaxInt32 or out-of-range TLEN, matching htslib, instead of truncating.

- record.go: Pos/PNext/TLen -> int64; EndPosition() -> int64.
- sam_reader.go: ParseInt(...,64) for POS/PNEXT/TLEN. sam_writer.go already
  formatted via int64/FormatInt -> output byte-identical for small values.
- bam_reader.go: sign-extend on-disk int32 into the int64 fields (decodeInto +
  decodeShallowInto). bam_writer.go: range guards (clear error beyond int32).
- CRAM: convert at the int32 series boundaries; AP/NP/TS encoding byte-exact for
  normal inputs.
- Call-site fixups (behavior-preserving int32 casts where a downstream field is
  32-bit): samtools fixmate/markdup/phase*/stats, bcftools mpileup, bedtobam.
- TestHtslibLongRefs t.Skip removed (runs vs upstream, POS~1e10);
  TestBAMWriterLongPositionErrors added; ~15 test files updated for int32->int64
  Record literals.

Also converts the upstream-build/obtain-failure guards in the touched bcftools/
samtools parity tests from t.Fatalf to t.Skipf (CI hermetic test job: building
upstream is an env condition, not a test failure; the upstream-parity job, with
htslib pre-built, still enforces).

Validation: go build/vet/gofmt clean; go test ./... + go test -race on touched
packages clean; parity-pipeline -scale=small DIVERGE=0; TestHtslibLongRefs PASS.

+++ pkg/htsgo/alnio/threaded_test.go
@@ -34,7 +34,7 @@ func buildMultiBlockBAM(t *testing.T, n int) []byte {
 	for i := 0; i < n; i++ {
 		rec := &sam.Record{
 			QName: fmt.Sprintf("read%06d", i),
-			Flag:  0, RName: "chr1", Pos: int32((i*37)%99999000 + 1), MapQ: 60,
+			Flag:  0, RName: "chr1", Pos: int64((i*37)%99999000 + 1), MapQ: 60,
 			Cigar: cig, Seq: string(seq), Qual: qual,
 			RNext: "*", PNext: 0, TLen: 0,
 		}
